Admin moze potvrdit i odbit zahtjev za dvoranu
Sada se zapravo ispravno kreiraju dvorana mjesto adresa objekti u tablicama u bazi i povezu kako spada.

+++ source_code/backend/dvoranko-backend/src/main/java/fer/leprogi/dvoranko/dto/ZahtjevOglasDTO.java
@@ -16,7 +16,7 @@ public class ZahtjevOglasDTO {
     private Integer kapacitet;
 //    private Iterable<KategorijaDTO> kategorije;
 
-    private Integer postalCode;
+    private Long postalCode;
     private String city;
     private String street;
     private String streetNumber;
